+++ examples/years_fmt.rs
@@ -19,26 +19,25 @@ fn main() {
         (UnixTime::from_year(2022), 0), //To complete our histogram, we manually specify when 2021 ends.
     ];
 
-    let xc = UnixTime::ctx_fmt(|d| {
-        write!(
-            d.writer,
-            "Time between {} and {} in {}",
-            d.min, d.max, d.step
-        )
-    })
-    .marker(UnixTime::from_year(2025))
-    .with_fmt(|mut t| {
-        t.val.default_tick_fmt(&mut t.writer, t.step, t.info)?;
-        write!(t.writer, " yr")
-    });
 
-    let mut s = poloto::Plotter::new(
-        "Number of Wikipedia Articles",
-        xc,
-        i128::ctx("Number of articles").no_dash().marker(0),
+    //knowldge of canvas dim
+    let canvas=poloto::Canvas::new();
+
+    //compute min and max
+    let mut data=poloto::plot().line("foo",&data).find_bounds();
+
+    //compute step info
+    let ticks=canvas.gen_ticks(&data);
+
+
+    println!(
+        "{}<style>{}{}</style>{}{}",
+        poloto::SVG_HEADER,
+        poloto::STYLE_CONFIG_DARK_DEFAULT,
+        ".poloto_line{stroke-dasharray:2;stroke-width:1;}",
+        poloto::disp(|w| canvas.render(w,data,ticks,())),
+        poloto::SVG_END
     )
-    .histogram("", &data)
-    .move_into();
 
-    println!("{}", poloto::disp(|a| s.simple_theme(a)));
 }
+
