+++ src/render/mod.rs
@@ -21,6 +21,90 @@ struct ScaleData<X: PlotNum, Y: PlotNum> {
     aspect_offset: f64,
 }
 
+pub fn line_fill<T: std::fmt::Write>(
+    path: &mut tagger::PathBuilder<T>,
+    mut it: impl Iterator<Item = [f64; 2]>,
+    base_line: f64,
+    add_start_end_base: bool,
+) -> fmt::Result {
+    if let Some([startx, starty]) = it.next() {
+        use tagger::PathCommand::*;
+
+        let mut last = [startx, starty];
+        let mut last_finite = None;
+        let mut first = true;
+        for [newx, newy] in it {
+            match (
+                newx.is_finite() && newy.is_finite(),
+                last[0].is_finite() && last[1].is_finite(),
+            ) {
+                (true, true) => {
+                    if first {
+                        if add_start_end_base {
+                            path.put(M(last[0], base_line))?;
+                            path.put(L(last[0], last[1]))?;
+                        } else {
+                            path.put(M(last[0], last[1]))?;
+                        }
+                        first = false;
+                    }
+                    last_finite = Some([newx, newy]);
+                    path.put(L(newx, newy))?;
+                }
+                (true, false) => {
+                    path.put(M(newx, newy))?;
+                }
+                (false, true) => {
+                    path.put(L(last[0], base_line))?;
+                }
+                _ => {}
+            };
+            last = [newx, newy];
+        }
+        if let Some([x, _]) = last_finite {
+            if add_start_end_base {
+                path.put(L(x, base_line))?;
+            }
+            path.put(Z(""))?;
+        }
+    }
+    Ok(())
+}
+
+pub fn line<T: std::fmt::Write>(
+    path: &mut tagger::PathBuilder<T>,
+    mut it: impl Iterator<Item = [f64; 2]>,
+) -> fmt::Result {
+    if let Some([startx, starty]) = it.next() {
+        use tagger::PathCommand::*;
+
+        let mut last = [startx, starty];
+        let mut first = true;
+        for [newx, newy] in it {
+            match (
+                newx.is_finite() && newy.is_finite(),
+                last[0].is_finite() && last[1].is_finite(),
+            ) {
+                (true, true) => {
+                    if first {
+                        path.put(M(last[0], last[1]))?;
+                        first = false;
+                    }
+                    path.put(L(newx, newy))?;
+                }
+                (true, false) => {
+                    path.put(M(newx, newy))?;
+                }
+                _ => {}
+            };
+            last = [newx, newy];
+        }
+    }
+    Ok(())
+}
+
+
+/*
 //Returns error if the user supplied format functions don't work.
 //Panics if the element tag writing writes fail
 pub fn render<X: PlotNum, Y: PlotNum, T: std::fmt::Write>(
@@ -62,272 +146,6 @@ pub fn render<X: PlotNum, Y: PlotNum, T: std::fmt::Write>(
 
     let spacing = padding / 3.0;
     let legendx1 = width - padding / 1.2 + padding / 30.0;
-
-    let mut color_iter = {
-        let max = if let Some(nn) = plotter.num_css_classes {
-            nn
-        } else {
-            usize::MAX
-        };
-
-        (0..max).cycle()
-    };
-
-    for (i, mut p) in plotter.plots.drain(..).enumerate() {
-        let legendy1 = paddingy - padding / 8.0 + (i as f64) * spacing;
-
-        let name_exists = writer
-            .elem("text", |d| {
-                d.attr("class", "poloto_text poloto_legend_text")?;
-                d.attr("alignment-baseline", "middle")?;
-                d.attr("text-anchor", "start")?;
-                d.attr("font-size", "large")?;
-                d.attr("x", width - padding / 1.2)?;
-                d.attr("y", paddingy + (i as f64) * spacing)
-            })?
-            .build(|d| {
-                let mut wc = num::WriteCounter::new(d.writer_safe());
-                p.plots.write_name(&mut wc)?;
-                Ok(wc.get_counter() != 0)
-            })?;
-
-        let aa = xcontext.scale(minx, [minx, maxx], scalex2);
-        let bb = ycontext.scale(miny, [miny, maxy], scaley2);
-
-        struct PlotIter<X, Y> {
-            basex_ii: f64,
-            basey_ii: f64,
-            rangex_ii: [X; 2],
-            rangey_ii: [Y; 2],
-            maxx_ii: f64,
-            maxy_ii: f64,
-        }
-        impl<X: PlotNum, Y: PlotNum> PlotIter<X, Y> {
-            fn gen_iter<'a>(
-                &'a self,
-                p: &'a mut Plot<X, Y>,
-                xcontext: &'a mut dyn PlotNumContext<Num = X>,
-                ycontext: &'a mut dyn PlotNumContext<Num = Y>,
-            ) -> impl Iterator<Item = [f64; 2]> + 'a {
-                p.plots.iter_second().map(move |(x, y)| {
-                    [
-                        self.basex_ii + xcontext.scale(x, self.rangex_ii, self.maxx_ii),
-                        self.basey_ii - ycontext.scale(y, self.rangey_ii, self.maxy_ii),
-                    ]
-                })
-            }
-        }
-
-        let plot_iter = PlotIter {
-            basex_ii: aspect_offset + padding - aa,
-            basey_ii: height - paddingy + bb,
-            rangex_ii: [minx, maxx],
-            rangey_ii: [miny, maxy],
-            maxx_ii: scalex2,
-            maxy_ii: scaley2,
-        };
-
-        let colori = color_iter.next().unwrap();
-
-        match p.plot_type {
-            PlotType::Line => {
-                if name_exists {
-                    writer.single("line", |d| {
-                        d.attr(
-                            "class",
-                            format_args!(
-                                "poloto_line poloto_legend_icon poloto{}stroke poloto{}legend",
-                                colori, colori
-                            ),
-                        )?;
-                        d.attr("stroke", "black")?;
-                        d.attr("x1", legendx1)?;
-                        d.attr("x2", legendx1 + padding / 3.0)?;
-                        d.attr("y1", legendy1)?;
-                        d.attr("y2", legendy1)
-                    })?;
-                }
-
-                writer.single("path", |d| {
-                    d.attr("class", format_args!("poloto_line poloto{}stroke", colori))?;
-                    d.attr("fill", "none")?;
-                    d.attr("stroke", "black")?;
-                    d.path(|a| line(a, plot_iter.gen_iter(&mut p, xcontext, ycontext)))
-                })?;
-            }
-            PlotType::Scatter => {
-                if name_exists {
-                    writer.single("line", |d| {
-                        d.attr(
-                            "class",
-                            format_args!(
-                                "poloto_scatter poloto_legend_icon poloto{}stroke poloto{}legend",
-                                colori, colori
-                            ),
-                        )?;
-                        d.attr("stroke", "black")?;
-                        d.attr("x1", legendx1 + padding / 30.0)?;
-                        d.attr("x2", legendx1 + padding / 30.0)?;
-                        d.attr("y1", legendy1)?;
-                        d.attr("y2", legendy1)
-                    })?;
-                }
-
-                writer.single("path", |d| {
-                    d.attr(
-                        "class",
-                        format_args!("poloto_scatter poloto{}stroke", colori),
-                    )?;
-                    d.path(|a| {
-                        use tagger::PathCommand::*;
-                        for [x, y] in plot_iter
-                            .gen_iter(&mut p, xcontext, ycontext)
-                            .filter(|&[x, y]| x.is_finite() && y.is_finite())
-                        {
-                            a.put(M(x, y))?;
-                            a.put(H_(0))?;
-                        }
-                        Ok(())
-                    })
-                })?;
-            }
-            PlotType::Histo => {
-                if name_exists {
-                    writer.single("rect", |d| {
-                        d.attr(
-                            "class",
-                            format_args!(
-                                "poloto_histo poloto_legend_icon poloto{}fill poloto{}legend",
-                                colori, colori
-                            ),
-                        )?;
-                        d.attr("x", legendx1)?;
-                        d.attr("y", legendy1 - padding / 30.0)?;
-                        d.attr("width", padding / 3.0)?;
-                        d.attr("height", padding / 20.0)?;
-                        d.attr("rx", padding / 30.0)?;
-                        d.attr("ry", padding / 30.0)
-                    })?;
-                }
-
-                writer
-                    .elem("g", |d| {
-                        d.attr("class", format_args!("poloto_histo poloto{}fill", colori))
-                    })?
-                    .build(|writer| {
-                        let mut last = None;
-                        //TODO dont necesarily filter?
-                        for [x, y] in plot_iter
-                            .gen_iter(&mut p, xcontext, ycontext)
-                            .filter(|&[x, y]| x.is_finite() && y.is_finite())
-                        {
-                            if let Some((lx, ly)) = last {
-                                writer.single("rect", |d| {
-                                    d.attr("x", lx)?;
-                                    d.attr("y", ly)?;
-                                    d.attr(
-                                        "width",
-                                        (padding * 0.02).max((x - lx) - (padding * 0.02)),
-                                    )?;
-                                    d.attr("height", height - paddingy - ly)
-                                })?;
-                            }
-                            last = Some((x, y))
-                        }
-                        Ok(())
-                    })?;
-            }
-            PlotType::LineFill => {
-                if name_exists {
-                    writer.single("rect", |d| {
-                        d.attr(
-                            "class",
-                            format_args!(
-                                "poloto_linefill poloto_legend_icon poloto{}fill poloto{}legend",
-                                colori, colori
-                            ),
-                        )?;
-                        d.attr("x", legendx1)?;
-                        d.attr("y", legendy1 - padding / 30.0)?;
-                        d.attr("width", padding / 3.0)?;
-                        d.attr("height", padding / 20.0)?;
-                        d.attr("rx", padding / 30.0)?;
-                        d.attr("ry", padding / 30.0)
-                    })?;
-                }
-
-                writer.single("path", |d| {
-                    d.attr(
-                        "class",
-                        format_args!("poloto_linefill poloto{}fill", colori),
-                    )?;
-                    d.path(|path| {
-                        line_fill(
-                            path,
-                            plot_iter.gen_iter(&mut p, xcontext, ycontext),
-                            height - paddingy,
-                            true,
-                        )
-                    })
-                })?;
-            }
-            PlotType::LineFillRaw => {
-                if name_exists {
-                    writer.single("rect", |d| {
-                        d.attr(
-                            "class",
-                            format_args!(
-                                "poloto_linefillraw poloto_legend_icon poloto{}fill poloto{}legend",
-                                colori, colori
-                            ),
-                        )?;
-                        d.attr("x", legendx1)?;
-                        d.attr("y", legendy1 - padding / 30.0)?;
-                        d.attr("width", padding / 3.0)?;
-                        d.attr("height", padding / 20.0)?;
-                        d.attr("rx", padding / 30.0)?;
-                        d.attr("ry", padding / 30.0)
-                    })?;
-                }
-
-                writer.single("path", |d| {
-                    d.attr(
-                        "class",
-                        format_args!("poloto_linefillraw poloto{}fill", colori),
-                    )?;
-                    d.path(|path| {
-                        line_fill(
-                            path,
-                            plot_iter.gen_iter(&mut p, xcontext, ycontext),
-                            height - paddingy,
-                            false,
-                        )
-                    })
-                })?;
-            }
-        }
-    }
-
-    draw_base(
-        plotter,
-        &mut writer,
-        DrawData {
-            width: crate::WIDTH as f64,
-            height: crate::HEIGHT as f64,
-            padding: 150.0,
-            paddingy: 100.0,
-        },
-        ScaleData {
-            minx,
-            maxx,
-            miny,
-            maxy,
-            scalex: scalex2,
-            scaley: scaley2,
-            preserve_aspect,
-            aspect_offset,
-        },
-    )?;
-
-    Ok(())
 }
+
+*/
